requirements.txt
requests
python-dotenv
pyllamacpp